+++ CHANGELOG.md
@@ -1,5 +1,4 @@
 
-
 # Changelog
 All notable changes to this project will be documented in this file.
 
@@ -12,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Misc improvements to the DQE
 - Fix Project Creation UI issue
 - Fix issue with whitespace confusing encryption key paths
+- Add "View Parent Tree" functionality
 - Add ability to add description to new cohort versions
 - Add functionality to revert to historical cohort version
 - Fix issue with YAML mode object load ordering